capi_image_builder/requirements.txt
gitpython
semver
pytest
pylint
openstacksdk
